pyproject.toml
[tool.poetry]
name = "log-sorter"
version = "0.1.0"
description = ""
authors = ["artbutakov <art.butakov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
sqlalchemy = "^2.0.29"
wtforms = "^3.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
